Move migrated viewsets into dedicated viewsets.py modules
Third-stage cleanup after the wagtail-modeladmin migration: relocate the
viewset classes out of wagtail_hooks.py into sibling viewsets.py modules
(following the existing rca/users/viewsets.py convention), leaving only the
hook registrations in wagtail_hooks.py. Behaviour-preserving move only.

- rca/utils/viewsets.py: TaxonomyViewSet base, the 22 taxonomy ModelViewSets,
  ProgrammeStudyModeIndexView, and TaxonomiesViewSetGroup. wagtail_hooks.py
  keeps the register_admin_viewset hook and the unrelated rich-text external
  link handler.
- rca/enquire_to_study/viewsets.py: DisableCreatePermissionPolicy,
  EnquiryFormSubmissionFilterSet, EnquiryFormSubmissionIndexView, and
  EnquiryFormSubmissionViewSet. wagtail_hooks.py keeps the
  register_admin_viewset and register_admin_urls hooks.
- Update the two test imports to the new modules.

Full suite passes (245 tests).
